requirements.txt
#

# These requirements were autogenerated by pipenv

# To regenerate from the project's Pipfile, run:

#

#    pipenv lock --requirements

#



-i https://pypi.org/simple

asgiref==3.4.1; python_version >= '3.6'

coverage==6.2

crispy-bootstrap5==0.6

django-ckeditor==6.2.0

django-crispy-forms==1.13.0; python_version >= '3.6'

django-js-asset==1.2.2

django-phonenumber-field==6.0.0

django-quill-editor==0.1.34

django

mysqlclient==2.1.0

pillow==9.0.0

sqlparse==0.4.2; python_version >= '3.5'

tzdata==2021.5; sys_platform == 'win32'

